Major refactor to separate concerns and reduce the size of files.
