+++ app/analysis_API.py
@@ -1,7 +1,35 @@
-import json
+"""
+Functions that make up the analysis API.
+"""
 from app import app_object
-
 from . import database
+import json
+
+
+@app_object.route("/get_parametric_path/", methods=["POST"])
+def get_parametric_path():
+    """
+    A list of observation IDs, along with an instrumentation point ID, will be given in the request body.
+    """
+    data = json.loads(request.get_data())
+    observation_ids = data["observation_ids"]
+    instrumentation_point_id = data["instrumentation_point_id"]
+
+    print("getting intersection of paths up to observations with IDs %s, based on instrumentation point with ID %i" % \
+          (str(observation_ids), instrumentation_point_id))
+
+    intersection_data = database.compute_intersection(observation_ids, instrumentation_point_id)
+
+    return json.dumps(intersection_data)
+
+
+@app_object.route("/get_path_condition_sequence/<observation_id>/", methods=["GET"])
+def get_path_condition_sequence(observation_id):
+    """
+    Given an observation ID, determine the sequence of path conditions leading to it through the SCFG with which it's associated.
+    """
+    print("getting condition sequence")
+    return json.dumps(database.compute_condition_sequence_and_path_length(observation_id))
 
 
 @app_object.route("/client/list_functions_2")
